refactor(navigation): extract NavigationSyncModifier (#295)
Fifth slice of Phase 2a. ContentView still carried six onChange
handlers (~60 lines) reconciling SwiftUI @State (layerSelection /
phaseSelection / storedLayerIndex / storedPhaseIndex) with
ContentViewModel's @Published ground-truth IDs and indices. These
form two parallel state machines for the same conceptual values; the
handlers translate between them on every external change.

Bundles all six into a NavigationSyncModifier + a .navigationSync(...)
View extension. The synchronization noise lives in one focused file
with a doc-comment that enumerates the six observer paths and what
each one writes. ContentView's contentWithEvents now reads as
genuine lifecycle (.task / .ignoresSafeArea / sync-status handler)
without the index-conversion churn.

Also removes the now-unused private adjustPhaseSelection() helper —
its single caller was one of the moved onChange handlers, and the
modifier inlines the same logic.

ContentView drops 372 → 313 lines (-59). NavigationSyncModifier is
119 lines, comfortably under the Phase 2a ≤200 target.

Phase 2b (#296) will refactor PhaseNavigator and the layer filtering
system, at which point this modifier collapses into a proper
NavigationViewModel that owns the @State directly. The current
modifier shape is a deliberate stepping stone.
